Reworked .travis.yml and tox.ini; fixed flake8 in setup.py; added author in setup script; set to be only one universal EV
