+++ .travis.yml
@@ -3,11 +3,7 @@ language: python
 python:
 - '3.5'
 - '3.6'
-matrix:
-  include:
-  - python: 3.7
-    dist: xenial
-    sudo: true
+- '3.7'
 
 install:
   - pip install -r tests/requirements.txt
@@ -18,9 +14,9 @@ script: tox
 
 after_success:
   - wget https://raw.githubusercontent.com/DiscordHooks/travis-ci-discord-webhook/master/send.sh
-  - chmod +x send.sh
+  - chmod u+x send.sh
   - ./send.sh success $WEBHOOK_URL
 after_failure:
   - wget https://raw.githubusercontent.com/DiscordHooks/travis-ci-discord-webhook/master/send.sh
-  - chmod +x send.sh
+  - chmod u+x send.sh
   - ./send.sh failure $WEBHOOK_URL
