+++ tests/.env.example
@@ -1,1 +1,1 @@
-webhook_url = https://discordapp.com/api/webhooks/XXX/XXX`
+WEBHOOK_URL = https://discordapp.com/api/webhooks/XXX/XXX
