+++ tox.ini
@@ -5,8 +5,7 @@ exclude = __init__.py
 envlist = py35,py36,py37
 
 [testenv]
+deps = python-dotenv
 changedir = tests
 sitepackages = true
-whitelist_externals = pytest
-commands = pytest
-passenv = webhook_url
+commands = python -m unittest
